requirements.txt
PyPDF2==3.0.1
Pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.0.76
tqdm==4.66.1 